pyproject.toml
[project]
name = "dst-py"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Kamaal Farah", email = "kamaal.f1@gmail.com" }
]
dependencies = [
    "fastapi[standard]>=0.115.6",
    "requests>=2.32.3",
    "sqlmodel>=0.0.22",
    "email-validator>=2.2.0",
    "pyjwt>=2.10.1",
    "pydantic-settings>=2.7.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.36",
    "bcrypt>=4.2.1",
    "pydantic>=2.10.4",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.8.4",
    "pre-commit>=4.0.1",
    "pytest>=8.3.4",
    "pytest-xdist>=3.6.1",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.1",
    "freezegun>=1.5.1",
    "responses>=0.25.3",
    "types-requests>=2.32.0.20241016",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/dst_py"]
